app finished

+++ app.js
@@ -1,22 +1,30 @@
 const addBox = document.querySelector(".add-box")
 const popupBox = document.querySelector(".popup-box")
-const closeIcon = document.querySelector(".uil-times")
-const addButton = document.querySelector("button")
-const titleTag = document.querySelector("input")
-const descTag = document.querySelector("textarea")
+const popupTitle = popupBox.querySelector("header p")
+const closeIcon = popupBox.querySelector(".uil-times")
+const addButton = popupBox.querySelector("button")
+const titleTag = popupBox.querySelector("input")
+const descTag = popupBox.querySelector("textarea")
 
 const months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"]
 
 const notes = JSON.parse(localStorage.getItem("notes") || "[]")
+let isUpdate = false, updateId;
+
 addBox.addEventListener("click", () => {
+    titleTag.focus();
     popupBox.classList.add("show")
 })
 
 closeIcon.addEventListener("click", () => {
+    isUpdate = false
     popupBox.classList.remove("show")
-    titleTag.value = ""
-    descTag.value = ""
-})
+    titleTag.value = "";
+    descTag.value = "";
+    addButton.innerText = "Add Note"
+    popupTitle.innerText = "Add a new Note"
+}
+)
 
 
 function showNotes() {
@@ -32,7 +40,7 @@ function showNotes() {
             <div class="settings">
                 <i onclick="showMenu(this)" class="uil uil-ellipsis-h"></i>
                 <ul class="menu">
-                    <li><i class="uil uil-pen">Edit</i></li>
+                    <li onclick="updateNote(${index}, '${note.title}', '${note.description}')"><i class="uil uil-pen">Edit</i></li>
                     <li onclick="deleteNote(${index})" ><i class="uil uil-trash">Delete</i></li>
                 </ul>
             </div>
@@ -54,6 +62,8 @@ function showMenu (elem) {
 }
 
 function deleteNote (noteId) {
+    let confirmDel = confirm("Are you sure you want to delete this note ?")
+    if(!confirmDel) return;
     notes.splice(noteId,1) // removing selected node from array/tasks
     // saving update notes to localstorage
     localStorage.setItem("notes",JSON.stringify(notes));
@@ -78,6 +88,13 @@ addButton.addEventListener("click", e => {
             description: noteDesc,
             date: `${day} ${month} ${year}`
         }
+        if(!isUpdate){
+            notes.push(noteInfo)
+        }
+        else {
+            isUpdate = false
+            notes[updateId] = noteInfo; // updating specified note
+        }
         console.log(noteInfo);
         notes.push(noteInfo); // adding new note to notes
         // saving notes to local storage
@@ -89,3 +106,14 @@ addButton.addEventListener("click", e => {
     
    
 })
+
+
+function updateNote (noteId,title,desc) {
+        isUpdate = true
+        updateId = noteId
+        addBox.click(); // open addbox to update note
+        addButton.innerText = "Update Note"
+        popupTitle.innerText = "Update a Note"
+        titleTag.value = title
+        descTag.value  = desc
+}
